aiserver/test/QAMath/requirements.txt
# QA 搜索系统依赖
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.28.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0

# 向量数据库 (选择一个)
faiss-cpu>=1.7.4
# faiss-gpu  # 如果有 GPU 可用

